+++ src/Mailables/AbstractMailable.php
@@ -8,6 +8,8 @@ use Illuminate\Queue\SerializesModels;
 
 abstract class AbstractMailable extends BaseMailable
 {
+    // todo: refactor to allow interfaces to set properties without passing to constructor
+    // todo: refactor to use basic mailable building instead of views (new class?)
     use Queueable, SerializesModels;
 
     /**
@@ -64,6 +66,7 @@ abstract class AbstractMailable extends BaseMailable
         $this->call_to_action = $call_to_action ?? $this->call_to_action;
         $this->view = $view ?? $this->view;
 
+        // todo: use config values
         $this->onQueue('mail');
         $this->onConnection('database');
     }